server: add `prices` list to subscription schemas
Had to refactor customer schemas into two separate modules to avoid circular imports.
