+++ server/polar/license_key/schemas.py
@@ -5,7 +5,7 @@ from dateutil.relativedelta import relativedelta
 from pydantic import UUID4, AliasChoices, AliasPath, Field
 
 from polar.benefit.schemas import BenefitID
-from polar.customer.schemas import CustomerBase
+from polar.customer.schemas.customer import CustomerBase
 from polar.exceptions import ResourceNotFound, Unauthorized
 from polar.kit.schemas import Schema
 from polar.kit.utils import generate_uuid, utc_now
